+++ module/class/actions_dolicatalog.class.php
@@ -243,7 +243,7 @@ class ActionsDoliCatalog
 		$out .= '</div>';
 
 		$out .= '<script type="application/json" id="dolicatalog-config">'.json_encode($config).'</script>';
-		$out .= '<script src="'.dol_buildpath('/dolicatalog/js/dolicatalog.js', 1).'?v='.urlencode(dolicatalogGetVersion()).'"></script>';
+		$out .= '<script src="'.dol_buildpath('/dolicatalog/js/dolicatalog.js', 1).'?v='.urlencode(dolicatalogAssetVersion('/dolicatalog/js/dolicatalog.js')).'"></script>';
 		if ($wrap === 'tablerow') {
 			$out .= '</td></tr>';
 		}